Support for loading template defaults from plist

+++ PSPrintSDK/OLPrintOrder.m
@@ -112,7 +112,7 @@ static id stringOrEmptyString(NSString *str) {
     }
     
     if (self.currenciesSupported.count == 0) {
-        NSAssert(NO, @"currenciesSupported.count == 0. There are %d jobs that are part of this order", self.jobs.count);
+        NSAssert(NO, @"currenciesSupported.count == 0. There are %lu jobs that are part of this order", (unsigned long) self.jobs.count);
     }
     
     code = self.currenciesSupported[0]; // return the first currency supported if the user hasn't specified one explicitly
